+++ core/src/main/scala/com/krux/hyperion/cli/Action.scala
@@ -1,0 +1,7 @@
+package com.krux.hyperion.cli
+
+import com.krux.hyperion.DataPipelineDef
+
+private[hyperion] trait Action {
+  def apply(options: Options, pipelineDef: DataPipelineDef): Boolean
+}