+++ core/src/main/scala/com/krux/hyperion/cli/AwsAction.scala
@@ -1,0 +1,12 @@
+package com.krux.hyperion.cli
+
+import com.krux.hyperion.{HyperionAwsClient, DataPipelineDef}
+
+private[hyperion] trait AwsAction extends Action {
+
+  def apply(options: Options, client: HyperionAwsClient): Boolean
+
+  def apply(options: Options, pipelineDef: DataPipelineDef): Boolean =
+    apply(options, HyperionAwsClient(pipelineDef, options.region, options.roleArn))
+
+}