+++ core/src/main/scala/com/krux/hyperion/resource/SparkCluster.scala
@@ -5,7 +5,7 @@ import org.slf4j.LoggerFactory
 import com.krux.hyperion.adt.HType._
 import com.krux.hyperion.adt.{HInt, HDuration, HDouble, HString, HBoolean}
 import com.krux.hyperion.aws.AdpEmrCluster
-import com.krux.hyperion.common.PipelineObjectId
+import com.krux.hyperion.common.{HttpProxy, PipelineObject, PipelineObjectId}
 import com.krux.hyperion.HyperionContext
 
 /**
@@ -13,8 +13,8 @@ import com.krux.hyperion.HyperionContext
  */
 class SparkCluster private (
   val id: PipelineObjectId,
-  val sparkVersion: HString,
-  val amiVersion: HString,
+  val sparkVersion: Option[HString],
+  val amiVersion: Option[HString],
   val supportedProducts: Option[HString],
   val standardBootstrapAction: Seq[HString],
   val bootstrapAction: Seq[HString],
@@ -43,7 +43,11 @@ class SparkCluster private (
   val initTimeout: Option[HDuration],
   val terminateAfter: Option[HDuration],
   val actionOnResourceFailure: Option[ActionOnResourceFailure],
-  val actionOnTaskFailure: Option[ActionOnTaskFailure]
+  val actionOnTaskFailure: Option[ActionOnTaskFailure],
+  val httpProxy: Option[HttpProxy],
+  val releaseLabel: Option[HString],
+  val applications: Seq[HString],
+  val configuration: Option[EmrConfiguration]
 ) extends EmrCluster {
 
   val logger = LoggerFactory.getLogger(SparkCluster.getClass)
@@ -58,8 +62,8 @@ class SparkCluster private (
   })
 
   def copy(id: PipelineObjectId = id,
-    sparkVersion: HString = sparkVersion,
-    amiVersion: HString = amiVersion,
+    sparkVersion: Option[HString] = sparkVersion,
+    amiVersion: Option[HString] = amiVersion,
     supportedProducts: Option[HString] = supportedProducts,
     standardBootstrapAction: Seq[HString] = standardBootstrapAction,
     bootstrapAction: Seq[HString] = bootstrapAction,
@@ -88,20 +92,24 @@ class SparkCluster private (
     initTimeout: Option[HDuration] = initTimeout,
     terminateAfter: Option[HDuration] = terminateAfter,
     actionOnResourceFailure: Option[ActionOnResourceFailure] = actionOnResourceFailure,
-    actionOnTaskFailure: Option[ActionOnTaskFailure] = actionOnTaskFailure
+    actionOnTaskFailure: Option[ActionOnTaskFailure] = actionOnTaskFailure,
+    httpProxy: Option[HttpProxy] = httpProxy,
+    releaseLabel: Option[HString] = releaseLabel,
+    applications: Seq[HString] = applications,
+    configuration: Option[EmrConfiguration] = configuration
   ) = new SparkCluster(id, sparkVersion, amiVersion, supportedProducts, standardBootstrapAction, bootstrapAction,
     enableDebugging, hadoopSchedulerType, keyPair, masterInstanceBidPrice, masterInstanceType,
     coreInstanceBidPrice, coreInstanceCount, coreInstanceType,
     taskInstanceBidPrice, taskInstanceCount, taskInstanceType, region, availabilityZone, resourceRole, role, subnetId,
     masterSecurityGroupId, additionalMasterSecurityGroupIds, slaveSecurityGroupId, additionalSlaveSecurityGroupIds,
     useOnDemandOnLastAttempt, visibleToAllUsers, initTimeout, terminateAfter,
-    actionOnResourceFailure, actionOnTaskFailure)
+    actionOnResourceFailure, actionOnTaskFailure, httpProxy, releaseLabel, applications, configuration)
 
   def named(name: String) = this.copy(id = id.named(name))
   def groupedBy(group: String) = this.copy(id = id.groupedBy(group))
 
-  def withSparkVersion(sparkVersion: HString) = this.copy(sparkVersion = sparkVersion)
-  def withAmiVersion(ver: HString) = this.copy(amiVersion = ver)
+  def withSparkVersion(sparkVersion: HString) = this.copy(sparkVersion = Option(sparkVersion))
+  def withAmiVersion(ver: HString) = this.copy(amiVersion = Option(ver))
   def withSupportedProducts(products: HString) = this.copy(supportedProducts = Option(products))
   def withBootstrapAction(action: HString*) = this.copy(bootstrapAction = bootstrapAction ++ action)
   def withDebuggingEnabled() = this.copy(enableDebugging = Option(HBoolean.True))
@@ -130,15 +138,21 @@ class SparkCluster private (
   def terminatingAfter(terminateAfter: HDuration) = this.copy(terminateAfter = Option(terminateAfter))
   def withActionOnResourceFailure(actionOnResourceFailure: ActionOnResourceFailure) = this.copy(actionOnResourceFailure = Option(actionOnResourceFailure))
   def withActionOnTaskFailure(actionOnTaskFailure: ActionOnTaskFailure) = this.copy(actionOnTaskFailure = Option(actionOnTaskFailure))
+  def withHttpProxy(proxy: HttpProxy) = this.copy(httpProxy = Option(proxy))
+  def withReleaseLabel(releaseLabel: HString) = this.copy(releaseLabel = Option(releaseLabel), amiVersion = None)
+  def withApplication(application: HString*) = this.copy(applications = this.applications ++ application)
+  def withConfiguration(configuration: EmrConfiguration) = this.copy(configuration = Option(configuration))
 
   lazy val instanceCount = coreInstanceCount + taskInstanceCount + 1
+
+  override def objects: Iterable[PipelineObject] = configuration.toList ++ httpProxy.toList
 
   lazy val serialize = new AdpEmrCluster(
     id = id,
     name = id.toOption,
-    amiVersion = Option(amiVersion.serialize),
+    amiVersion = amiVersion.map(_.serialize),
     supportedProducts = supportedProducts.map(_.serialize),
-    bootstrapAction = Seq(s"s3://support.elasticmapreduce/spark/install-spark,-v,$sparkVersion,-x")
+    bootstrapAction = Seq(s"s3://support.elasticmapreduce/spark/install-spark,-v,${sparkVersion.get},-x")
       ++ standardBootstrapAction.map(_.serialize) ++ bootstrapAction.map(_.serialize),
     enableDebugging = enableDebugging.map(_.serialize),
     hadoopSchedulerType = hadoopSchedulerType.map(_.serialize),
@@ -171,14 +185,22 @@ class SparkCluster private (
     initTimeout = initTimeout.map(_.serialize),
     terminateAfter = terminateAfter.map(_.serialize),
     actionOnResourceFailure = actionOnResourceFailure.map(_.serialize),
-    actionOnTaskFailure = actionOnTaskFailure.map(_.serialize)
+    actionOnTaskFailure = actionOnTaskFailure.map(_.serialize),
+    httpProxy = httpProxy.map(_.ref),
+    releaseLabel = releaseLabel.map(_.serialize),
+    applications = applications.map(_.serialize),
+    configuration = configuration.map(_.ref)
   )
 
 }
 
 object SparkCluster {
 
-  def apply()(implicit hc: HyperionContext) = new SparkCluster(
+  def apply()(implicit hc: HyperionContext): SparkCluster = apply(None)
+
+  def apply(configuration: EmrConfiguration)(implicit hc: HyperionContext): SparkCluster = apply(Option(configuration))
+
+  private def apply(configuration: Option[EmrConfiguration])(implicit hc: HyperionContext): SparkCluster = new SparkCluster(
     id = PipelineObjectId(SparkCluster.getClass),
     sparkVersion = hc.emrSparkVersion,
     amiVersion = hc.emrAmiVersion,
@@ -211,7 +233,11 @@ object SparkCluster {
     initTimeout = None,
     terminateAfter = hc.emrTerminateAfter.map(x => x: HDuration),
     actionOnResourceFailure = None,
-    actionOnTaskFailure = None
+    actionOnTaskFailure = None,
+    httpProxy = None,
+    releaseLabel = hc.emrReleaseLabel,
+    applications = Seq.empty,
+    configuration = configuration
   )
 
 }